feat(LG-50): 유저 도메인을 수정한다
