+++ src/main/java/com/linkgem/domain/user/UserReader.java
@@ -1,0 +1,8 @@
+package com.linkgem.domain.user;
+
+import java.util.Optional;
+
+public interface UserReader {
+
+    Optional<User> find(Long id);
+}
