+++ src/main/java/com/linkgem/infrastructure/user/UserRepository.java
@@ -1,0 +1,8 @@
+package com.linkgem.infrastructure.user;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.linkgem.domain.user.User;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+}
